+++ packages/web/app/components/EditInvoiceModal/index.jsx
@@ -1,63 +1,17 @@
 import React, { useCallback, useEffect, useState } from 'react';
 import styled from 'styled-components';
-import { Divider, Grid, IconButton, Menu } from '@material-ui/core';
-import { MoreVert } from '@material-ui/icons';
+import { Divider } from '@material-ui/core';
 import { Modal } from '../Modal';
 import { TranslatedEnum, TranslatedText } from '../Translation';
-import { AutocompleteField, DateField, Field, Form } from '../Field';
-import { useApi, useSuggester } from '../../api';
+import { Form } from '../Field';
+import { useApi } from '../../api';
 import { Colors } from '../../constants';
 import { FormSubmitCancelRow } from '../ButtonRow';
 import { DataFetchingTable } from '../Table';
 import { DateDisplay } from '../DateDisplay';
 import { INVOICE_LINE_TYPE_LABELS } from '@tamanu/constants';
 import { Button } from '../Button';
-import { DeleteItemModal } from './DeleteItemModal';
-
-const PriceText = styled.span`
-  margin-right: 16px;
-`;
-
-const StyledItemRow = styled(Grid)`
-  font-size: 11px;
-  padding-left: 10px;
-  padding-right: 10px;
-  padding-bottom: 4px;
-  background: white;
-  border-left: 1px solid ${Colors.outline};
-  border-right: 1px solid ${Colors.outline};
-  border-top: 1px solid ${Colors.outline};
-  ${props => props.hasBorderBottom &&
-    `border-bottom: 1px solid ${Colors.outline}; padding-bottom: 0px;`}
-`;
-
-const StyledItemHeader = styled(Grid)`
-  padding: 14px;
-  font-weight: 500;
-  border-radius: 4px 4px 0 0;
-  border: 1px solid ${Colors.outline};
-`;
-
-const KebabMenuItem = styled.div`
-  width: 124px;
-  font-weight: 400;
-  font-size: 11px;
-  line-height: 15px;
-  padding-left: 8px;
-  padding-right: 8px;
-  cursor: pointer;
-  border-radius: 4px;
-  padding-bottom: 8px;
-  :hover {
-    background: rgba(255, 255, 255, 0.15);
-  }
-`;
-
-const StyledMenu = styled(Menu)`
-  & .MuiList-padding {
-    padding-bottom: 0;
-  }
-`;
+import { ItemHeader, ItemRow } from './ItemRow';
 
 const LinkText = styled.div`
   font-weight: 500;
@@ -67,21 +21,6 @@ const LinkText = styled.div`
   margin-top: 10px;
   margin-bottom: 10px;
   cursor: pointer;
-`;
-
-const StyledIconButton = styled(IconButton)`
-  margin-bottom: 4px;
-  margin-left: auto;
-`;
-
-const PriceCell = styled.div`
-  margin-left: 10%;
-  display: flex;
-  align-items: center;
-`;
-
-const ItemCodeText = styled.div`
-  margin-left: 5%;
 `;
 
 const SingleAddButton = styled(Button)`
@@ -129,185 +68,6 @@ const StyledDivider = styled(Divider)`
   margin: 26px -40px 32px -40px;
 `;
 
-const ItemHeader = () => {
-  return <StyledItemHeader container alignItems='center' spacing={1}>
-    <Grid item xs={2}>
-      <TranslatedText
-        stringId="general.date.label"
-        fallback="Date"
-      />
-    </Grid>
-    <Grid item xs={4}>
-      <TranslatedText
-        stringId="invoice.modal.addInvoice.details.label"
-        fallback="Details"
-      />
-    </Grid>
-    <Grid item xs={1}>
-      <ItemCodeText>
-        <TranslatedText
-          stringId="invoice.table.column.code"
-          fallback="Code"
-        />
-      </ItemCodeText>
-    </Grid>
-    <Grid item xs={3}>
-      <TranslatedText
-        stringId="invoice.modal.addInvoice.orderedBy.label"
-        fallback="Ordered by"
-      />
-    </Grid>
-    <Grid item xs={2}>
-      <PriceCell>
-        <TranslatedText
-          stringId="invoice.modal.addInvoice.price.label"
-          fallback="Price"
-        />
-      </PriceCell>
-    </Grid>
-  </StyledItemHeader>
-};
-
-const ACTION_MODALS = {
-  ADD_DISCOUNT: 'addDiscount',
-  ADD_MARKUP: 'addMarkup',
-  DELETE: 'delete',
-}
-
-const ItemRow = ({ index, hasBorderBottom, category, onDelete, rowData: defaultRowData, isDeleteDisabled }) => {
-  const invoiceLineTypeSuggester = useSuggester('invoiceLineTypes');
-  const practitionerSuggester = useSuggester('practitioner');
-  const [actionModal, setActionModal] = useState('');
-  const [anchorEl, setAnchorEl] = useState(null);
-  const open = Boolean(anchorEl);
-  const [rowData, setRowData] = useState({
-    code: defaultRowData?.type
-      ? <TranslatedEnum
-        prefix="invoice.line.property.type"
-        value={defaultRowData.type}
-        enumValues={INVOICE_LINE_TYPE_LABELS}
-      /> : "",
-    details: defaultRowData?.name || "",
-    date: defaultRowData?.date || "",
-    orderedBy: defaultRowData?.orderedBy || "",
-    price: defaultRowData?.price || "",
-  });
-
-  const onOpenKebabMenu = (event) => {
-    setAnchorEl(event.currentTarget);
-  };
-
-  const handleCloseKebabMenu = () => {
-    setAnchorEl(null);
-  };
-
-  const updateRowData = ({ itemType, price }) => {
-    setRowData(prevRowData => ({
-      ...prevRowData,
-      code: <TranslatedEnum
-        prefix="invoice.line.property.type"
-        value={itemType}
-        enumValues={INVOICE_LINE_TYPE_LABELS}
-      />,
-      price,
-    }))
-  };
-
-  const handleActionModal = value => setActionModal(value);
-
-  const handleDeleteItem = () => {
-    onDelete();
-    handleActionModal('')
-  };
-
-  return <StyledItemRow container alignItems='center' spacing={1} hasBorderBottom={hasBorderBottom}>
-    <Grid item xs={2}>
-      <Field
-        name={"date_" + index}
-        required
-        component={DateField}
-        saveDateAsString
-        size="small"
-      />
-    </Grid>
-    <Grid item xs={4}>
-      <Field
-        name={"invoiceLineTypeId_" + index}
-        required
-        component={AutocompleteField}
-        suggester={invoiceLineTypeSuggester}
-        customCallback={data => updateRowData(data)}
-        size="small"
-      />
-    </Grid>
-    <Grid item justifyContent='center' xs={1}>
-      <ItemCodeText>
-        {category || rowData.code}
-      </ItemCodeText>
-    </Grid>
-    <Grid item xs={3}>
-      <Field
-        name={"orderedById_" + index}
-        required
-        component={AutocompleteField}
-        suggester={practitionerSuggester}
-        size="small"
-      />
-    </Grid>
-    <Grid item xs={2}>
-      <PriceCell>
-        <PriceText>
-          {rowData.price}
-        </PriceText>
-        <StyledIconButton
-          onClick={onOpenKebabMenu}
-        >
-          <MoreVert />
-        </StyledIconButton>
-      </PriceCell>
-      <StyledMenu
-        anchorEl={anchorEl}
-        getContentAnchorEl={null}
-        open={open}
-        onClose={handleCloseKebabMenu}
-        anchorOrigin={{
-          vertical: 'bottom',
-          horizontal: 'right',
-        }}
-        transformOrigin={{
-          vertical: 'top',
-          horizontal: 'right',
-        }}
-      >
-        <KebabMenuItem onClick={() => handleActionModal(ACTION_MODALS.ADD_DISCOUNT)}>
-          <TranslatedText
-            stringId="invoice.modal.editInvoice.addDiscount"
-            fallback="Add discount"
-          />
-        </KebabMenuItem>
-        <KebabMenuItem onClick={() => handleActionModal(ACTION_MODALS.ADD_MARKUP)}>
-          <TranslatedText
-            stringId="invoice.modal.editInvoice.addMarkup"
-            fallback="Add markup"
-          />
-        </KebabMenuItem>
-        <KebabMenuItem onClick={() => !isDeleteDisabled && handleActionModal(ACTION_MODALS.DELETE)}>
-          <TranslatedText
-            stringId="invoice.modal.editInvoice.delete"
-            fallback="Delete"
-          />
-        </KebabMenuItem>
-      </StyledMenu>
-    </Grid>
-    <DeleteItemModal
-      open={actionModal === ACTION_MODALS.DELETE}
-      onClose={() => handleActionModal('')}
-      onDelete={handleDeleteItem}
-      lineItems={rowData}
-    />
-  </StyledItemRow>
-};
-
 export const EditInvoiceModal = ({ open, onClose, invoiceId, displayId, invoiceLineItems }) => {
   const defaultRowList = invoiceLineItems.length ? invoiceLineItems.map(item => ({
     invoiceLineTypeId: item.invoiceLineTypeId,
@@ -321,7 +81,7 @@ export const EditInvoiceModal = ({ open, onClose, invoiceId, displayId, invoiceL
   const api = useApi();
 
   const handleAddRow = (rowData) => {
-    let newRowList = [...rowList];
+    const newRowList = [...rowList];
     if (Array.isArray(rowData) && rowData.length) {
       rowData.forEach(newItem => {
         const idExists = newRowList.some(item => item && item.id === newItem.id);
